Fixed: AidaFileReader.hh(20): warning C4244 in MSVC: 'return' : conversion from 'const unsigned __int64' to 'unsigned int', possible loss of data

+++ main/include/eudaq/AidaFileReader.hh
@@ -17,7 +17,7 @@ namespace eudaq {
     virtual ~AidaFileReader();
     bool readNext();
     std::string Filename() const { return m_filename; }
-    unsigned RunNumber() const { return m_runNumber; }
+    unsigned long long RunNumber() const { return m_runNumber; }
     std::string getJsonConfig() { return m_json_config; }
     std::string getJsonPacketInfo();
     std::shared_ptr<eudaq::AidaPacket> GetPacket() const { return m_packet; };
